pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envelop"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = "MIT"
keywords = []
authors = [
  { name = "Lucino772", email = "lucapalmi772@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pydantic==2.6.4",
  "grpcio[protobuf]==1.62.1"
]

[project.urls]
Documentation = "https://github.com/Lucino772/envelop#readme"
Issues = "https://github.com/Lucino772/envelop/issues"
Source = "https://github.com/Lucino772/envelop"

[tool.hatch.version]
path = "src/envelop/__about__.py"

[tool.hatch.envs.grpc]
dependencies = [
  "grpcio-tools==1.62.1"
]

[tool.hatch.envs.grpc.scripts]
build = """\
python -m grpc_tools.protoc \
  -Ienvelop/services/proto=protos \
  --python_out=./src \
  --pyi_out=./src \
  --grpc_python_out=./src \
  protos/**.proto
"""

[tool.ruff]
line-length = 88
extend-exclude = ["*_pb2.py", "*_pb2.pyi","*_pb2_grpc.py" ]
